Updated OsmSharp.

+++ src/Itinero.IO.Osm/Streams/RouterDbStreamTarget.cs
@@ -118,7 +118,7 @@ namespace Itinero.IO.Osm.Streams
                         { // node is not a vertex.
                             return Constants.NO_VERTEX;
                         }
-                        return this.AddCoreNode(node.Id.Value, node.Latitude.Value, node.Longitude.Value);
+                        return this.AddCoreNode(node.Id.Value, (float)node.Latitude.Value, (float)node.Longitude.Value);
                     }));
                 }
             }
@@ -142,7 +142,7 @@ namespace Itinero.IO.Osm.Streams
                     { // node is not a vertex.
                         return Constants.NO_VERTEX;
                     }
-                    return this.AddCoreNode(node.Id.Value, node.Latitude.Value, node.Longitude.Value);
+                    return this.AddCoreNode(node.Id.Value, (float)node.Latitude.Value, (float)node.Longitude.Value);
                 },
                 (vehicleType, sequence) =>
                 {
@@ -315,7 +315,7 @@ namespace Itinero.IO.Osm.Streams
                 var index = _nodeIndex.TryGetIndex(node.Id.Value);
                 if (index != long.MaxValue)
                 { // node is a routing node, store it's coordinates.
-                    _nodeIndex.SetIndex(index, node.Latitude.Value, node.Longitude.Value);
+                    _nodeIndex.SetIndex(index, (float)node.Latitude.Value, (float)node.Longitude.Value);
                 }
             }
         }
